// src/payload-types.ts
export interface User {
  id: string;
  username: string;
  avatar?: (string | null) | Media;
  bio?: string | null;
  followers?: (string | User)[] | null;
  following?: (string | User)[] | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  type: 'image' | 'video';
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: string;
  title?: string | null;
  caption: string;
  video: string | Media;
  author: string | User;
  comments?: (string | Comment)[] | null;
  likes?: (string | User)[] | null;
  /**
   * Add hashtags for your post (without the # symbol)
   */
  hashtags?:
    | {
        value: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "comments".
 */
export interface Comment {
  id: string;
  content: string;
  author: string | User;
  post: string | Post;
  updatedAt: string;
  createdAt: string;
}